[#30]: Fix warnings

+++ engine/Game/Player.cpp
@@ -139,7 +139,7 @@ Player::GetReloadTime() const
 void
 Player::ChangeWepon(int32_t idx)
 {
-   m_currentWeapon = m_weapons.at(idx).get();
+   m_currentWeapon = m_weapons.at(static_cast<size_t>(idx)).get();
 }
 
 int32_t
